+++ pkg/heroku/account_feature.go
@@ -49,7 +49,7 @@ func (c *Client) AccountFeatureInfo(accountFeatureIdentity string) (*AccountFeat
 // lr is an optional ListRange that sets the Range options for the paginated
 // list of results.
 func (c *Client) AccountFeatureList(lr *ListRange) ([]AccountFeature, error) {
-	req, err := c.NewRequest("GET", "/account/features", nil)
+	req, err := c.NewRequest("GET", "/account/features", nil, nil)
 	if err != nil {
 		return nil, err
 	}