+++ pkg/heroku/release.go
@@ -52,7 +52,7 @@ func (c *Client) ReleaseInfo(appIdentity string, releaseIdentity string) (*Relea
 // appIdentity is the unique identifier of the Release's App. lr is an optional
 // ListRange that sets the Range options for the paginated list of results.
 func (c *Client) ReleaseList(appIdentity string, lr *ListRange) ([]Release, error) {
-	req, err := c.NewRequest("GET", "/apps/"+appIdentity+"/releases", nil)
+	req, err := c.NewRequest("GET", "/apps/"+appIdentity+"/releases", nil, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -94,12 +94,13 @@ type ReleaseCreateOpts struct {
 //
 // appIdentity is the unique identifier of the Release's App. release is the
 // unique identifier of release.
-func (c *Client) ReleaseRollback(appIdentity string, release string) (*Release, error) {
+func (c *Client) ReleaseRollback(appIdentity, release, message string) (*Release, error) {
 	params := struct {
 		Release string `json:"release"`
 	}{
 		Release: release,
 	}
+	rh := RequestHeaders{CommitMessage: message}
 	var releaseRes Release
-	return &releaseRes, c.Post(&releaseRes, "/apps/"+appIdentity+"/releases", params)
+	return &releaseRes, c.PostWithHeaders(&releaseRes, "/apps/"+appIdentity+"/releases", params, rh.Headers())
 }